chemical_visualizer/backend/requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.14
pandas>=2.0
django-cors-headers>=4.0
reportlab>=4.0
